return callbacks to main bot.py file
